Add --keep-position
When playing very large playlists (i.e. consisting of hundreds of
entries each with several hours long), it is rather unpleasant to lose
the playback position. Even with --save-position-on-quit, this can
still occur due to mpv exiting abnormally, e.g. due to a power
failure. In such circumstances, it is still much more preferable to
resume from the last saved position instead of completely losing it.

For this purpose, add --keep-position, which asks mpv to not erase the
position file on start-up.

DRAFT! CURRENTLY THIS DOESN'T WORK FOR PLAYLISTS because of how mpv
finds which file to resume playback in a playlist from. TODO: To make
this work, we need to make mpv clean up the state file when moving to
another playlist file (and probably immediately write the playback
position), or rethink how mpv saves/finds playback position for
playlists.

+++ player/configfiles.c
@@ -451,7 +451,8 @@ void mp_load_playback_resume(struct MPContext *mpctx, const char *file)
         MP_INFO(mpctx, "Resuming playback. This behavior can "
                "be disabled with --no-resume-playback.\n");
         try_load_config(mpctx, fname, M_SETOPT_PRESERVE_CMDLINE, MSGL_V);
-        unlink(fname);
+        if (!mpctx->opts->position_keep)
+            unlink(fname);
     }
     talloc_free(fname);
 }
